Renamed a few packages to internal in the org.eclipse.smarthome.storage.mapdb bundle. See openhab/static-code-analysis#100
