+++ bundles/storage/org.eclipse.smarthome.storage.mapdb/src/main/java/org/eclipse/smarthome/storage/mapdb/internal/MapDbStorage.java
@@ -5,7 +5,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.eclipse.smarthome.storage.mapdb;
+package org.eclipse.smarthome.storage.mapdb.internal;
 
 import java.util.ArrayList;
 import java.util.Collection;
